- SettingsDetailViewController -> DetailSettingsViewController 네이밍 수정
